requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery
redis
django-celery-beat
channels
channels-redis
django-guardian
daphne
whitenoise